+++ app/Http/Middleware/HandleInertiaRequests.php
@@ -42,6 +42,11 @@ class HandleInertiaRequests extends Middleware
         return [
             ...parent::share($request),
             'name' => config('app.name'),
+            'locale' => app()->getLocale(),
+            'translations' => [
+                'posts' => trans('posts'),
+                'common' => trans('common'),
+            ],
             'quote' => ['message' => trim($message), 'author' => trim($author)],
             'auth' => [
                 'user' => $request->user(),