+++ resources/js/types/index.d.ts
@@ -25,6 +25,11 @@ export interface NavItem {
 
 export interface SharedData {
     name: string;
+    locale: string;
+    translations: {
+        posts: Record<string, string>;
+        common: Record<string, string>;
+    };
     quote: { message: string; author: string };
     auth: Auth;
     ziggy: Config & { location: string };