+++ resources/js/types/shared.d.ts
@@ -3,6 +3,11 @@ import type { Auth } from './index';
 
 declare interface SharedData {
     name: string;
+    locale: string;
+    translations: {
+        posts: Record<string, string>;
+        common: Record<string, string>;
+    };
     quote: { message: string; author: string };
     auth: Auth;
     ziggy: Config & { location: string };